+++ drivers/usb/dwc3/io.h
@@ -1,3 +1,4 @@
+/* SPDX-License-Identifier: GPL-2.0 */
 /**
  * io.h - DesignWare USB3 DRD IO Header
  *
@@ -10,8 +11,6 @@
  * to uboot.
  *
  * commit 2c4cbe6e5a : usb: dwc3: add tracepoints to aid debugging
- *
- * SPDX-License-Identifier:     GPL-2.0
  *
  */
 
@@ -48,7 +47,7 @@ static inline void dwc3_writel(void __iomem *base, u32 offset, u32 value)
 	writel(value, base + offs);
 }
 
-static inline void dwc3_flush_cache(int addr, int length)
+static inline void dwc3_flush_cache(uintptr_t addr, int length)
 {
 	flush_dcache_range(addr, addr + ROUND(length, CACHELINE_SIZE));
 }
