+++ include/dwc3-uboot.h
@@ -1,10 +1,9 @@
+/* SPDX-License-Identifier: GPL-2.0 */
 /* include/dwc3-uboot.h
  *
  * Copyright (c) 2015 Texas Instruments Incorporated - http://www.ti.com
  *
  * Designware SuperSpeed USB uboot init
- *
- * SPDX-License-Identifier:     GPL-2.0
  */
 
 #ifndef __DWC3_UBOOT_H_
@@ -13,7 +12,7 @@
 #include <linux/usb/otg.h>
 
 struct dwc3_device {
-	int base;
+	unsigned long base;
 	enum usb_dr_mode dr_mode;
 	u32 maximum_speed;
 	unsigned tx_fifo_resize:1;
